backend/services/sibill-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.5.2
sqlalchemy==2.0.23
databases==0.8.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
openai==1.55.3
